requirements.txt
   python-dotenv==1.1.0
   pymysql==1.1.2
   mysql-replication==1.0.12
